Move TransformProperty to rendering/debug.dart to fix cyclic deps issue.
