+++ packages/flutter/lib/src/rendering/layer.dart
@@ -10,6 +10,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/painting.dart';
 import 'package:vector_math/vector_math_64.dart';
 
+import 'debug.dart';
 import 'node.dart';
 
 /// A composited layer.
